+++ resources/views/components/code.blade.php
@@ -1,3 +1,0 @@
-<div {{ $attributes }}>
-	<img class="mx-auto" src="{{ $qrcode }}" alt="{{ __('QR Code') }}" />
-</div>