+++ resources/views/errors/access-denied.blade.php
@@ -1,5 +1,0 @@
-@extends('errors.layout')
-
-@section('title', __('Access Denied'))
-@section('code', '403')
-@section('message', __($exception->getMessage() ?: 'Access Denied'))