+++ resources/views/errors/expired.blade.php
@@ -1,5 +1,0 @@
-@extends('errors.layout')
-
-@section('title', __('Expired'))
-@section('code', '401')
-@section('message', __($exception->getMessage() ?: 'Expired'))