+++ resources/views/errors/invalid-code.blade.php
@@ -1,5 +1,0 @@
-@extends('errors.layout')
-
-@section('title', __('Invalid code'))
-@section('code', '422')
-@section('message', __($exception->getMessage() ?: 'Invalid code'))