+++ src/Exceptions/InvitationAlreadyAcceptedException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TwentySixB\LaravelInvitations\Exceptions;
+
+use Exception;
+
+class InvitationAlreadyAcceptedException extends Exception
+{
+}