+++ src/Exceptions/InvitationAlreadyRejectedException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TwentySixB\LaravelInvitations\Exceptions;
+
+use Exception;
+
+class InvitationAlreadyRejectedException extends Exception
+{
+}