+++ src/InvitationsServiceProvider.php
@@ -2,14 +2,11 @@
 
 namespace TwentySixB\LaravelInvitations;
 
-use Livewire\Livewire;
+use Illuminate\Support\Facades\Gate;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 use Spatie\LaravelPackageTools\Package;
-use Illuminate\Support\Facades\Blade;
+use TwentySixB\LaravelInvitations\Console\Commands\DispatchExpiredInvitations;
 use TwentySixB\LaravelInvitations\Console\Commands\PurgeExpiredInvitations;
-use TwentySixB\LaravelInvitations\Livewire\Lister;
-use TwentySixB\LaravelInvitations\Livewire\Viewer;
-use TwentySixB\LaravelInvitations\Models\Invitation;
 use TwentySixB\LaravelInvitations\Policies\InvitationPolicy;
 
 /**
@@ -18,16 +15,6 @@ use TwentySixB\LaravelInvitations\Policies\InvitationPolicy;
  */
 class InvitationsServiceProvider extends PackageServiceProvider
 {
-
-	/**
-     * The policy mappings for the application.
-     *
-     * @var array
-     */
-    protected $policies = [
-        Invitation::class => InvitationPolicy::class,
-    ];
-
     /**
      * @inheritDoc
      *
@@ -38,11 +25,14 @@ class InvitationsServiceProvider extends PackageServiceProvider
     {
         $package->name('laravel-invitations')
             ->hasConfigFile()
-			->hasMigration('create_invitations_table')
-			// FIXME: Submit issue, middleware doesnt work this way.
-            // ->hasRoute('web')
-			->hasCommand(PurgeExpiredInvitations::class)
-            ->hasViews('invitations');
+            ->hasMigrations([
+                'create_invitations_table',
+                'alter_invitations_table_add_accepted_and_rejected_at',
+            ])
+            ->hasCommands([
+                DispatchExpiredInvitations::class,
+                PurgeExpiredInvitations::class,
+            ]);
     }
 
     /**
@@ -52,13 +42,6 @@ class InvitationsServiceProvider extends PackageServiceProvider
      */
     public function packageBooted() : void
     {
-		Livewire::component('invitations.viewer', Viewer::class);
-        Livewire::component('invitations.lister', Lister::class);
-
-		Blade::componentNamespace('TwentySixB\\LaravelInvitations\\View\\Components', 'invitations');
-
-		// TODO: Update component.
-		//Livewire::component('invitations.inviter', InviteUsers::class);
-
+        Gate::policy(config('invitations.models.invitation'), InvitationPolicy::class);
     }
 }