+++ tests/Invitable.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace TwentySixB\LaravelInvitations\Tests;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Invitable extends Model
+{
+    public $incrementing = false;
+
+    public $timestamps = false;
+}